% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFSCfunctions-package.R
\docType{package}
\name{AFSCFunctions}
\alias{AFSCFunctions}
\alias{AFSCFunctions-package}
\title{AFSCFunctions}
\description{
The \pkg{AFSCFunctions} package provides a standardized set of
bespoke functions for use by AFSC FMA staff.
}
\author{
Jason E. Jannot (jason.jannot@noaa.gov)
}
\keyword{internal}
