% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ak_db_connect_function.R
\name{ak_db_connect}
\alias{ak_db_connect}
\title{Connect to AFSC database using \code{odbc}}
\usage{
ak_db_connect()
}
\value{
A connection to the AFSC database.
}
\description{
\code{ak_db_connect} creates a connection to the AFSC database.
}
\details{
This function takes no arguments.

\code{ak_db_connect} takes no arguements.  It will ask for your AFSC username
and password and then create and test a connection to the database.
}
\note{
The function uses \code{odbc} to connect to the database. Currently
only connects to AFSC Oracle database. Might be modified in the future to
include other databases, e.g., AKRO.
Note also that this will \emph{not} create a connection in Rstudio.
}
\examples{
\dontrun{
con1 <- ak_db_connect()
tst <- odbc::dbGetQuery(con1, "SELECT * FROM norpac.atl_haul
                                                 FETCH FIRST 10 ROWS ONLY")
nrow(tst)
}

}
\seealso{
\code{odbc::dbConnect, odbc::dbGetQuery()}
\code{rstudioapi::askForPassword()}
}
