% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{chk_missing_nms}
\alias{chk_missing_nms}
\title{Which Names Are Missing From Data.Frame}
\usage{
chk_missing_nms(df1, df2)
}
\arguments{
\item{df1}{first data frame}

\item{df2}{second data frame}
}
\value{
Returns character vector
}
\description{
A low level wrapper to check for differences in names between two data.frames.
}
\details{
Checks for names of columns in each data.frame that does not appear in the other data.frame.  Prints the missing column names for each data frame.
}
