% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{colord}
\alias{colord}
\title{Sort Columns of Data.Frame by Names of Another Data.Frame}
\usage{
colord(x, y, suppress = FALSE)
}
\arguments{
\item{x}{data.frame whose columns will be sorted.}

\item{y}{data.frame whose columns will be used as template for sorting.}

\item{suppress}{suppress the printing of column names? Default = FALSE}
}
\value{
Returns \code{x}, with columns sorted according to \code{y}.
}
\description{
Function to return a data.frame with columns ordered in the same manner as another data.frame, in prep for \code{rbind()}
}
\details{
A low level function, used in many of the functions in this package. Usage is to return a data.frame (\code{x}) whose columns are sorted according to the template data.frame columns (\code{y}).
}
\examples{
\dontrun{
df2 <- colord(df2, df1)
}

}
