% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_formats_constant.R
\docType{data}
\name{date_formats}
\alias{date_formats}
\title{Date Formats}
\format{
An object of class \code{character} of length 18.
}
\usage{
date_formats
}
\value{
Returns a character vector of date formats.
}
\description{
\code{date_formats} is a character string containing the many
possible formats that a date could take.  For use with \code{pkg-lubridate}
functions.
}
\details{
\code{date_formats}is a constant that should not be exported from NAMESPACE.
}
\keyword{datasets}
