% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_query.R
\name{db_query}
\alias{db_query}
\title{Query a database}
\usage{
db_query(query, dsn = "channel_afsc")
}
\arguments{
\item{query}{the character string}

\item{dsn}{the dsn as specified in your .Renviron file}
}
\value{
Returns the output of from you query, typically a table.
}
\description{
Run a query without checking your connection
}
\details{
Runs a database query. This function will automatically see if you have an open connect to the database, and if not,
will re-establish the connection.

Uses the credential stored in the .Renviron file. By default, this function assumes you're connecting to the NORPAC
database but the \code{dsn} argument can be used to specify another alias.

As a reminder, you can edit your \code{.Renviron} file via \code{usethis::edit_r_environ()}.
}
\examples{
\dontrun{
   db_query(query = paste0("SELECT * FROM norpac.atl_lov_vessel"))
}

}
\seealso{
\code{usethis::edit_r_environ()}
}
