% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dots2char}
\alias{dots2char}
\title{Creates Character Vector from ...}
\usage{
dots2char(...)
}
\arguments{
\item{...}{from a function.}
}
\value{
Returns a character vector
}
\description{
\code{dots2char} converts \code{...} into a character vector.
}
