% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dated_file_function.R
\name{find_dated_file}
\alias{find_dated_file}
\title{Find and (optionally) Import a File with a Date in the File Name}
\usage{
find_dated_file(dte, path, filename, imprt = TRUE, ...)
}
\arguments{
\item{dte}{A character vector (length = 1) of either the word "recent",
which will be used to find the file name with the most recent date OR
a date (YYYY-MM-DD) which will be used to find the file name that contains
the specified date.}

\item{path}{A character vector (length = 1) of the path to search for files.
Converted to \code{base::file.path()} and then passed to
\code{base::list.files(path = path)}.}

\item{filename}{A character vector (length = 1) specifying the file name or
parts thereof, including regular expressions, but excluding \code{dte}.
Passed to\code{base::list.files(pattern = filename)}.}

\item{imprt}{A logical (default = TRUE). Do you wish to import the file?
A \code{.csv} file will be imported using \code{utils::read.csv()},
an R data file (\code{.Rda, .Rdata}) will be imported using
\code{base::load()}.}

\item{...}{arguments passed to \code{utils::read.csv()} or \code{base::load()}.}
}
\value{
Minimally, prints the name of the file(s). If \code{impt == TRUE}, imports a file.
}
\description{
\code{find_dated_file} searches a path for a file with a particular date
in the file name, or the file name with the most recent date, and optionally,
imports the file.
}
\details{
\code{FUNCTION NAME} DESCRIPTION
}
\note{
Most recent in this case, means the file name that contains the most recent date. The dte argument, if not 'recent', must be in the format of YYYY-MM-DD. Imported files can either be of \code{.csv} or \code{.Rda, .Rdata}.
}
\examples{
\dontrun{
# print filenames of file(s) with most recent date
find_dated_file(dte = "recent", path = "path2file",
                filename = "filename_or_partialname", imprt = FALSE)

# import file with most recent date in file name
# path + filename + most recent date (determined by \code{find_dated_file()})
#  must lead to a *single* file name which can be imported.
#   Otherwise, error.
find_dated_file(dte = "recent", path = "path2file",
                   filename = "filename_or_partialname", imprt = TRUE)
}

}
\seealso{
\code{base::list.files()}, \code{is_date_fmt() (in R/helper_functions.R)}
}
