% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flxtbl_fxn.R
\name{flxtbl_fxn}
\alias{flxtbl_fxn}
\title{Format Long Complicated Tables for P. halibut Report}
\usage{
flxtbl_fxn(
  df = NULL,
  hdrs = NULL,
  themFUN = flextable::theme_zebra,
  pd = 1,
  pd_part = "all",
  aln = "center",
  aln_part = "all",
  bd_part = "header"
)
}
\arguments{
\item{df}{A data.frame to be formatted}

\item{hdrs}{a named list of headers for the table, e.g., \code{list( OlD.Hdr1 = "New Header", old_headr2 = "New Header 2")}}

\item{themFUN}{a flextable theme.  See \code{flextable} for options.}

\item{pd}{a padding value (0, 1) to add space between table entries.}

\item{pd_part}{which part of the table should be padded? one of 'all', 'body', 'header', 'footer'.}

\item{aln}{an alignment position (character),  one of 'left', 'right', 'center', 'justify'.}

\item{aln_part}{which part of the table should be aligned? one of 'all', 'body', 'header', 'footer'}

\item{bd_part}{which part of the table should appear in \emph{bold}? one of 'all', 'body', 'header', 'footer'}
}
\description{
A wrapper function to format data for the executive summary table that summarizes Pacific halibut bycatch in the U.S. west coast groundfish fisheries.  Typically, this output will be used in a MS Word document (\code{.docx}).
}
\details{
A wrapper around \code{flextable}.
}
\examples{
\dontrun{
table_flex <- flxtbl_fxn(df)
}

}
\seealso{
\code{flextable::flextable()}
}
