% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMAAnalystFunctions-package.R
\docType{package}
\name{FMAAnalystFunctions-package}
\alias{FMAAnalystFunctions}
\alias{FMAAnalystFunctions-package}
\title{FMAAnalystFunctions: A Collection of Utility Functions Used by AFSC FMA Analysts}
\description{
Utility functions used by AFSC FMA Analysts.
}
\author{
\strong{Maintainer}: Jason Jannot \email{jason.jannot@noaa.gov} (\href{https://orcid.org/0000-0002-9231-1337}{ORCID})

Authors:
\itemize{
  \item Geoffrey Mayhew \email{geoff.mayhew@noaa.gov}
}

}
\keyword{internal}
