% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_Analyst_Functions-package.R
\docType{package}
\name{FMAAnalystFunctions}
\alias{FMAAnalystFunctions}
\alias{FMAAnalystFunctions-package}
\title{FMA Analyst Functions}
\description{
The \pkg{FMA Analyst Functions} package provides a standardized set of
bespoke functions for use by AFSC FMA Analyst team.
}
\author{
Jason E. Jannot (jason.jannot@noaa.gov)
}
\keyword{internal}
