\name{fmaDemo}
\alias{fmaDemo}
\docType{data}
\title{Simulated Data for Factor Mixture Analysis}
\description{
    The dataset contains simulated responses on a 1-4 ordinal scale of 600 individuals to a
    two-factors psychometric questionnaire.
}
\usage{
data(fmaDemo)
}
\format{A \code{data.frame} object.}

\examples{\dontrun{
data(fmaDemo)
}}
\keyword{datasets}
