% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table_csv_function.R
\name{format_table_csv}
\alias{format_table_csv}
\title{Formats Tables for Output to \code{.csv} Files with a Caption}
\usage{
format_table_csv(tb, cp = NULL)
}
\arguments{
\item{tb}{The table to be formatted. (data.frame, required)}

\item{cp}{The caption to be added to the table.(optional, character vector of \code{length == 1})}
}
\value{
Returns a data.frame
}
\description{
Function formats table headers and adds a caption to the table for output to \code{.csv}. The caption appears in the top left cell and the column headers become part of the data.frame.
}
\examples{
\dontrun{
tab_with_cap <- format_table_csv(df_to_format, cp = caption_to_add)
}

}
\seealso{
\code{footn_fxn}
}
