% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_download.R
\name{gdrive_download}
\alias{gdrive_download}
\title{Download a File from the Shared Google Drive}
\usage{
gdrive_download(local_path, gdrive_dribble, ver = NULL)
}
\arguments{
\item{local_path}{the local path to the file you wish to download, including the directory (folder paths) and the
name of the file, which should match the name of the file on the Gdrive.}

\item{gdrive_dribble}{the \code{dribble} class object representing the folder in the Gdrive where your desired file
resides.}

\item{ver}{Optional, default is NULL. If specified as a length 1 number, will download the requested version of the
file.}
}
\value{
Returns a message telling whether the download was performed and if so, where the file was saved.
}
\description{
\code{gdrive_download()} is used to download the most recent version of a file in the Gdrive, or, if specified, an older
version of a file. Currently only works with \code{.rdata} and \code{.rds} type files. Nearly all other files we use should be
maintained the project's GitHub repository.
}
\examples{
\dontrun{
  gdrive_download(local_path, gdrive_dribble)
}
}
