% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_set_dribble.R
\name{gdrive_set_dribble}
\alias{gdrive_set_dribble}
\title{Designate a Gdrive Folder as a Target for Uploads and Downloads}
\usage{
gdrive_set_dribble(gdrive_path, shared_id = "Analytics")
}
\arguments{
\item{gdrive_path}{a character string of the folder path, with each folder name followed by \code{/}}

\item{shared_id}{an alias for the Shared Drive to connect to. The default, \code{Analytics}, is so far the only alias that is recognized.}
}
\value{
Returns a dribble class object
}
\description{
Given a folder path of the folder structure of the Shared Google Drive, this function will create a \code{dribble}
class object for use by key functions to interact with the drive, including \code{gdrive_upload()} and
\code{gdrive_download()}. By default, it searches the FMA Analytical Services Program shared drive.
}
\details{
If you do not know the folder path to the folder you want to interact with, run \code{gdrive_dir()} to get
the view the folder structure of the Share Google Drive and copy-paste the \code{path}
}
\examples{
\dontrun{
   gdrive_set_dribble(gdrive_path = 'Google Drive Test/')
}

}
