% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date_fmt_function.R
\name{is_date_fmt}
\alias{is_date_fmt}
\title{Check Date Format}
\usage{
is_date_fmt(dte, dtefmt)
}
\arguments{
\item{dte}{a date string}

\item{dtefmt}{a date format (see \code{strptime} for formats)}
}
\value{
Returns a logical.
}
\description{
\code{is_date_fmt()} checks a date against a format to insure date is in
the correct format. This function should correctly handle NAs.
}
\note{
Based on function found here:
https://gist.github.com/micstr/69a64fbd0f5635094a53?permalink_comment_id=3432683#gistcomment-3432683
}
\seealso{
\code{as.Date, Date, strptime}
}
