% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdat_function.R
\name{load_rdat}
\alias{load_rdat}
\title{Load and Rename Object in Single Step}
\usage{
load_rdat(fileName)
}
\arguments{
\item{fileName}{The path and filename to the object to be loaded}
}
\value{
Loads a object into .GlobalEnv and assigns a name
}
\description{
\code{load_rdat} loads an object and assigns a new name in a single step
}
\note{
\strong{BEWARE}: If multiple objects were entered under a single \code{save()} call, this
function will only retrieve the \strong{first} object.
}
\examples{
\dontrun{
d <- load_rdat("~/blah/somegood.Rda")
}

}
\seealso{
\code{load}, \code{new.env}
}
