% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_confidential_data_function.R
\name{mask_conf_data}
\alias{mask_conf_data}
\title{Mask Confidential Data within Strata}
\usage{
mask_conf_data(confs, df, sybl = NULL, rnd = FALSE, decv = NULL)
}
\arguments{
\item{confs}{A data.frame of confidential strata, usually from \code{stratify_4phlb}.}

\item{df}{A data.frame whose strata will be masked upon output.}

\item{sybl}{A character vector (\code{length == 1}) giving the replacement text for confidential data. Defaults to \code{*}.}

\item{rnd}{A logical indicating if the values should be rounded using \code{round_df}. Default = FALSE. See details.}

\item{decv}{Only applies if \code{rnd == TRUE}. An integer indicating the number of decimal places to round, passed to \code{base::round} via \code{round_df}.}

\item{...}{Optional arguments \code{prnt, sym} passed to \code{round_df}.  See \code{?round_df} for specifications.}
}
\value{
Returns a \code{data.frame} with confidential strata masked.
}
\description{
A low level convenience function to mask data within confidential strata.
}
\details{
\code{mask_conf_data} will mask confidential data in the strata identified in the \code{confs} data.frame, which most often will be included in the output of \code{stratify_4phlb}. \code{mask_conf_data} replaces all numeric variables (\code{is.numeric}) with an asterisk (default), or optionally, with user supplied character(s), and in the process, convert all values to character.  If you want to preserve the values as numeric, use \code{sybl = NA}. Alternatively, set \code{rnd == TRUE} and the corresponding \code{decv} and optionally \code{...} to round all the output for printing as character.

Note: this function does \emph{not} test for confidentaility - it assumes that the strata provided in \code{confs} are correctly specified.
}
\examples{
\dontrun{
masked_df <- mask_conf_data(conf_strata, conf_df, sybl = "CONFIDENTIAL DATA")
}

}
\seealso{
\code{test_conf} to test for confidential strata.
}
