% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{match_col_classes}
\alias{match_col_classes}
\title{Makes Column Classes of two Data.Frames Consistent Across Common Column Names}
\usage{
match_col_classes(df1, df2)
}
\arguments{
\item{df1}{data.frame whose columns classes will be used as reference.}

\item{df2}{data.frame with at least 1 column name that matches a column name in df1. For all column names that match, columns classes will be  modified to match corresponding column classes in df1.}
}
\value{
Returns \code{df2}, with columns classes modified to match classes of shared column names in df1.
}
\description{
\code{matchColClasses} identifies columns that share the same name,, uses the master data frame (df1) to identify the shared columns classes, and then re-assigns column classes in df2 to match df1.
}
\details{
A low level function, used to ensure proper joins between data.frames within functions.
}
\examples{
\dontrun{
df2 <- matchColClasses(df1, df2)
}

}
