% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{namefix}
\alias{namefix}
\title{Makes Column Names of Output Data.Frame Consistent}
\usage{
namefix(df)
}
\arguments{
\item{df}{data.frame whose columns names will be fixed.}
}
\value{
Returns \code{df}, with columns names fixed.
}
\description{
\code{namefix} fixes the column names for consistency across functions.
}
\details{
A low level function, used in many packages. Usage is to fix column names for consistency across functions.
}
\examples{
\dontrun{
names(df) <- namefix(df)
}

}
