% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{neisna}
\alias{neisna}
\title{Is Any NA Present?}
\usage{
neisna(x)
}
\arguments{
\item{x}{any R object that can be checked for NA}
}
\value{
Returns logical
}
\description{
A low level wrapper for \code{any(is.na(x))}.
}
\details{
Wrapper for \code{any(is.na(x))}.
}
\examples{
\dontrun{
neisna(x)
}

}
