% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{qdf}
\alias{qdf}
\title{Quick Data Frame}
\usage{
qdf(..., saf = FALSE)
}
\arguments{
\item{...}{data.frame, see details.}

\item{saf}{strings as factors? Default = FALSE}
}
\value{
Returns the \code{data.frame}, with strings as characters, mainly from piping in the tidyverse.
}
\description{
Function to return a data.frame mainly from piping
}
\details{
A low level function, used in many of the functions in this package. Usage is to return a data.frame without factors from piping in the tidyverse.
}
\examples{
\dontrun{
df <- df \%>\%
  filter(...) \%>\%
  mutate(...) \%>\%
  select(...) \%>\%
  qdf()
}

}
