% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{qdr}
\alias{qdr}
\title{Quick Data Read CSV}
\usage{
qdr(pth)
}
\arguments{
\item{pth}{path to a \code{.csv} file.}
}
\value{
Returns the \code{.csv}, headers and with strings as characters.
}
\description{
A simple wrapper function for \code{read.csv(file =..., header = TRUE, stringsAsFactors = FALSE)}
}
\examples{
\dontrun{
df <- qdr(pth = "/path/to/file/filename.csv")
}

}
