% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{round_df}
\alias{round_df}
\title{Round Numeric Columns of a data.frame}
\usage{
round_df(df, digits, prnt = FALSE, sym = "NA")
}
\arguments{
\item{df}{Data.frame with numeric columns for rounding.}

\item{digits}{the number of digits to round, passed to \code{base::round()}.}

\item{prnt}{Do you want to print these values in a table format? If \code{FALSE} (default), no trailing zeros will be printed.}

\item{sym}{A character vector of (length == 1) used as a symbol for any 'NA's.}
}
\value{
Returns a data.frame with numeric columns rounded to \code{digits}
}
\description{
Rounds only numeric (i.e., \code{is.double()== TRUE}) columns of a data.frame to the level specified in \code{digits}.
}
\details{
Only columns that satisfy \code{is.double() == TRUE} will be rounded.  Thus, integers are not rounded.
}
