% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{scol}
\alias{scol}
\title{A Helper Function That Sorts the Object Names}
\usage{
scol(x)
}
\arguments{
\item{x}{any object that has names.}
}
\value{
Names sorted alphabetically.
}
\description{
\code{scol} sorts and prints the column names in alphabetical order.
}
\examples{
\dontrun{
scol(x)
}

}
