% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortnTab_function.R
\name{shortnTab}
\alias{shortnTab}
\title{Shorten Tables for Printing}
\usage{
shortnTab(df)
}
\arguments{
\item{dat}{A data.frame that will be printed as a table with zeros at bottom that can be cut-off (i.e., zeros that are uninformative).}
}
\value{
Returns the range of rows (e.g., x:y) to be \strong{included} in the data.frame.
}
\description{
\code{shortnTab} cuts off the bottom portion of a table where all the values are == 0 which makes for easier printing/reporting.
}
\details{
The most common usage is in a length frequency table where the last \emph{n} rows are length bins with uninformative zeros.
}
\examples{
\dontrun{
shorter_table <- long_table[shortnTab(long_table), ]
}

}
