% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_partial_year_function.R
\name{subset_partial_year}
\alias{subset_partial_year}
\title{Subset Data to Partial Years}
\usage{
subset_partial_year(.df, .finalmonth = 12, .finalday = 31)
}
\arguments{
\item{.df}{A data.frame with at least one date field.}

\item{.finalmonth}{A numeric value (0-12) representing the final month to be \emph{included} in each partial year's data. Defaults to 12.}

\item{.finalday}{A numeric value (1-31) representing the final day of the month to be \emph{included} in each partial year's data. Defaults to 31.}
}
\value{
Returns a data.frame with each year subset to the appropriate portion of the year.
}
\description{
\code{subset_partial_year} subsets each year included in the input data set to a portion of the year specified.
}
