% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{uni}
\alias{uni}
\title{Order Unique Values}
\usage{
uni(.x)
}
\arguments{
\item{.x}{an R object with a class or a numeric, complex, character or logical vector.}
}
\value{
Returns a vector of ordered unique values.
}
\description{
\code{uni()} Pulls the unique values (using \code{unique()}) and orders them (using \code{sort()}).
}
