% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_csv_list}
\alias{write_csv_list}
\title{Write Multiple CSVs from a list}
\usage{
write_csv_list(liste, fnms, rnms = FALSE)
}
\arguments{
\item{liste}{A list whose elements will be written to a csv.}

\item{fnms}{A character vector of file names (which can include a path, if desired) for each csv, where \code{length(fnms) == length(liste)}.}

\item{rnms}{Optional logical specifying whether or not to add the \code{row.names} to the output file. Default = \code{FALSE}.}
}
\value{
Nothing.
}
\description{
\code{write_csv_list()} Takes a list and writes each element to a csv.
}
\examples{
\dontrun{
write_csv_list(liste = listofdf, fnms = vectoroffilenames)
}

}
